added login, logout for worker

+++ controller/dashboard.controller.js
@@ -52,7 +52,7 @@ async function worker(req, res) {
   const { worker_id } = req.body;
   try {
     const { rows } = await pool.query(
-      `SELECT (SELECT SUM(hours) 
+      `SELECT (SELECT SUM(hours::numeric) 
           FROM attendences 
           WHERE EXTRACT(MONTH FROM date) = EXTRACT(MONTH FROM NOW()) AND EXTRACT(YEAR FROM date) = EXTRACT(YEAR FROM NOW()) AND id = $1) AS total_work_hours;`,
       [worker_id]
